Np.array.reshape() takes an int or tuple of ints as its second parameter.
This function is about rearranging multidimensional arrays, so the "shape"
numbers are always supposed to be integers, because they're counts of
rows and columns.

In a couple of places we were clearly relying on the floor-integer behavior
of python2 integer division.  I feel confident that moving these instances
in reshape() to the double // explicit floor integer division is a safe
change and is in line with the original behavior of the script.

Making this change cleared up some integration test failures, so that
feels like good evidence that this was correct.