Replace map() with a list comprehension, where possible
In python3, map() returns a Map iterator object, instead of a list.
In most cases that's fine, but if you intend to do array-access operations
on the result, then you need to cast it to a list first.

A better way to handle this is with a list comprehension.  For example:

Don't:
result = [map(float, item) for item in some_2d_list]

Do:
result = [[float(x) for x in item] for item in some_2d_list]

Making this change fixed a breaking unit test.