Force a predictable order of results in the espresso partial DOS txt file reader.
The tests are rather terrifyingly dependent on the order of a list,
and this list seems to be dependent on how a directory's files are read.

I sorted the file list as it's loaded from the disk, to make sure that
we always get a repeatable order in the resulting data structure.

I also reordered the reference elements in the test, to match.